+++ semana9/astromatch/src/kiddos/Home/ProfileCard/styled.js
@@ -2,53 +2,53 @@ import styled, { keyframes, css } from 'styled-components'
 import { Box } from "@material-ui/core";
 
 const right = keyframes`
-  from {
-	  transform: translate(0) rotate(0);
-    opacity: 1;
-  }
+    from {
+	    transform: translate(0) rotate(0);
+        opacity: 1;
+    }
 
-  to {
-	  transform: translate(-200px) rotate(-20deg);
-    opacity: 0;
-  }
+    to {
+	    transform: translate(-200px) rotate(-20deg);
+        opacity: 0;
+    }
 `;
 
 const left = keyframes`
-  from {
-    opacity: 1;
-	  transform: translate(0) rotate(0);
-  }
+    from {
+        opacity: 1;
+	    transform: translate(0) rotate(0);
+    }
 
-  to {
-    opacity: 0;
-	  transform: translate(200px) rotate(20deg);
-  }
+    to {
+        opacity: 0;
+	    transform: translate(200px) rotate(20deg);
+    }
 `;
 
 export const Container = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin: 1em;
-  transition: 0.5s;
-  animation: ${(props) => {
-    if(props.swipeLeft){
-      return css`${left} 1.5s`
-    } else if(props.swipeRight) {
-      return css`${right} 1.5s`
-    }
-  }};
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    margin: 1em;
+    transition: 0.5s;
+    animation: ${(props) => {
+        if (props.swipeLeft) {
+            return css`${left} 1.5s`
+        } else if (props.swipeRight) {
+            return css`${right} 1.5s`
+        }
+    }};
 `
 export const BlurredBackground = styled(Box)`
-  background-image: url(${(props) => props.bgimage});
+    background-image: url(${(props) => props.bgimage});
 	filter: blur(10px);
-  margin: 0;
+    margin: 0;
 	height: 100%;
 	width: 100%;
-  background-clip: border-box;
-  background-position: center;
-  background-repeat: no-repeat;
-  position: absolute;
+    background-clip: border-box;
+    background-position: center;
+    background-repeat: no-repeat;
+    position: absolute;
 `
 
 export const ProfilePicture = styled.img`
@@ -59,21 +59,21 @@ export const ProfilePicture = styled.img`
 
 
 export const StyledBox = styled(Box)`
-  height: 420px;
-  width: 300px;
-  display: flex;
+    height: 400px;
+    width: 300px;
+    display: flex;
 	align-items: center;
 	position: relative;
-  overflow: hidden;
+    overflow: hidden;
 `
 
 export const InfoBox = styled(Box)`
 	position: absolute;
-  bottom: 0;
+    bottom: 0;
 	width: 90%;
 	color: white;
 	display: flex;
 	flex-direction: column;
 	z-index: 2;
-  cursor: default;
+    cursor: default;
 `
